+++ WordPress-Extra/ruleset.xml
@@ -122,7 +122,7 @@
 
 	<!-- Warn against overriding WP global variables.
 		 https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/26 -->
-	<rule ref="WordPress.Variables.GlobalVariables"/>
+	<rule ref="WordPress.WP.GlobalVariablesOverride"/>
 
 	<!-- Encourage the use of strict ( === and !== ) comparisons.
 		 https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/242 -->
