+++ WordPress/ruleset.xml
@@ -162,4 +162,14 @@
 		<severity>0</severity>
 	</rule>
 
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.Variables.GlobalVariables.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.Variables.GlobalVariables.OverrideProhibited">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
